Refactor async states and filesystem commands
Introduce a dedicated IAsyncState and move async state types out of types.hpp into lib/TKardunio/asyncStates.hpp and src/fileSystem/asyncStates.hpp. Centralize per-command async storage via an AnyState union and placement-new pointers in commandFunctions, replacing previous static state instances. Add and reorganize filesystem command implementations into src/fileSystem/commands/ (erase, findFile, recieve, store/freeSpace/printFiles renames) and extend FAT API with removeEntry, entry_address, entry_size and readData. Fix various type/cast issues (i16/u16, int loop counters), correct EEPROM/fileStore byte handling, update includes, and set PlatformIO build_flags to -std=gnu++17. Also apply small formatting and control-flow cleanups in main and other modules.

+++ lib/TKardunio/collection.cpp
@@ -1,6 +1,6 @@
 #include "collection.hpp"
 
-inline bool isNumber(int8_t currentDecimal) {
+inline bool isNumber(i16 currentDecimal) {
   return currentDecimal > -1 && currentDecimal < 10;
 }
 
@@ -14,11 +14,11 @@ const Fstr* parseU8(StrSlice str, MutRef<u8> value) {
     u16 number = 0;
     for(usize i = 0; i < str.len(); i++) {
         number *= 10;
-        u8 currentDecimal = (u8)(str[i] - '0');
+        i16 currentDecimal = (i16)(str[i] - '0');
         if(!isNumber(currentDecimal))
             return F("!!error!! string is not valid number");
 
-        number += currentDecimal;
+        number += (u16)currentDecimal;
     }
 
     if(number > UINT8_MAX)
